requirements.txt
streamlit
python-nmap
flask
flask-socketio
requests